Added MPC for nonlinear planar crane, solved with cvxpy
Note: This isn't quite how this would really work with a nonlinear system. We'd be predicting using the linear model at each step, but using measurements of the current state of nonlinear system. As implemented here, the model used to generate the "measurements" for the MPC solution is the linear one.